requirements.txt
annotated-types==0.7.0
DateTime==5.5
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.0
pytz==2025.2
setuptools==80.9.0
slack_sdk==3.35.0
typing-inspection==0.4.1
typing_extensions==4.14.0
zope.interface==7.2
